+++ samplesheets/configapps/bih_generic/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BihGenericConfig(AppConfig):
+    name = 'samplesheets.configapps.bih_generic'